requirements.txt
torch==2.2.1
transformers==4.44.0
tiktoken==0.6.0
flash-attn==2.5.7
pillow
gradio==4.44.0
decord